backend/requirement.txt
Django==4.0
djangorestframework
facebook-sdk==3.1.0
django-cors-headers